Document playlist fixtures location
