+++ .gitignore
@@ -1,0 +1,3 @@
+/vendor/
+/.phpunit.result.cache
+/.phpunit.cache/
