+++ server/worker_process_canais.php
@@ -65,6 +65,8 @@ const CHANNEL_PROGRESS_DOWNLOAD = 5;
 const PROGRESS_MAX = 99;
 const CHANNEL_BATCH_UPDATE = 50;
 
+require_once __DIR__ . '/includes/channel_processing_helpers.php';
+
 $timeoutEnv = getenv('IMPORTADOR_M3U_TIMEOUT');
 $streamTimeout = ($timeoutEnv !== false && is_numeric($timeoutEnv) && (int) $timeoutEnv > 0)
     ? (int) $timeoutEnv
@@ -87,20 +89,6 @@ function logInfo(string $message): void
     } else {
         error_log($line);
     }
-}
-
-function sanitizeMessage(string $message): string
-{
-    $trimmed = trim($message);
-    if (function_exists('mb_substr')) {
-        return mb_substr($trimmed, 0, 2000, 'UTF-8');
-    }
-    return substr($trimmed, 0, 2000);
-}
-
-function formatBrazilianNumber(int $value): string
-{
-    return number_format($value, 0, ',', '.');
 }
 
 function updateJob(PDO $adminPdo, int $jobId, array $fields): void
@@ -154,78 +142,6 @@ function fetchJob(PDO $adminPdo, int $jobId): ?array
     return $job ?: null;
 }
 
-function getStreamTypeByUrl(string $url): array
-{
-    if (stripos($url, '/movie/') !== false || stripos($url, '/series/') !== false) {
-        return ['type' => 0, 'category_type' => ''];
-    }
-
-    return ['type' => 1, 'category_type' => 'live'];
-}
-
-/**
- * @return \Generator<int, array{url: string, tvg_logo: string, group_title: string, tvg_name: string}>
- */
-function extractChannelEntries(string $filePath): \Generator
-{
-    $handle = fopen($filePath, 'r');
-    if ($handle === false) {
-        throw new RuntimeException('Não foi possível ler o ficheiro M3U.');
-    }
-
-    $currentInfo = [
-        'tvg_logo' => '',
-        'group_title' => 'Canais',
-        'tvg_name' => '',
-    ];
-
-    try {
-        while (($line = fgets($handle)) !== false) {
-            $line = trim($line);
-            if ($line === '') {
-                continue;
-            }
-
-            if (stripos($line, '#EXTINF:') === 0) {
-                preg_match('/tvg-logo="(.*?)"/', $line, $logoMatch);
-                $currentInfo['tvg_logo'] = $logoMatch[1] ?? '';
-
-                $groupTitle = 'Canais';
-                if (preg_match('/group-title="(.*?)"/', $line, $groupMatch)) {
-                    $groupTitle = trim($groupMatch[1]);
-                }
-
-                $title = '';
-                $pos = strpos($line, '",');
-                if ($pos !== false) {
-                    $title = trim(substr($line, $pos + 2));
-                }
-                if ($title === '') {
-                    $parts = explode(',', $line, 2);
-                    $title = trim($parts[1] ?? '');
-                }
-
-                $currentInfo['group_title'] = $groupTitle !== '' ? $groupTitle : 'Canais';
-                $currentInfo['tvg_name'] = $title !== '' ? $title : 'Sem Nome';
-                continue;
-            }
-
-            if (!filter_var($line, FILTER_VALIDATE_URL)) {
-                continue;
-            }
-
-            yield [
-                'url' => $line,
-                'tvg_logo' => $currentInfo['tvg_logo'] ?? '',
-                'group_title' => $currentInfo['group_title'] ?? 'Canais',
-                'tvg_name' => $currentInfo['tvg_name'] ?? 'Sem Nome',
-            ];
-        }
-    } finally {
-        fclose($handle);
-    }
-}
-
 function getCategoryId(PDO $pdo, string $categoryName, string $categoryType): int
 {
     static $cache = [];
@@ -259,35 +175,6 @@ function getCategoryId(PDO $pdo, string $categoryName, string $categoryType): in
     $cache[$cacheKey] = $lastId;
 
     return $lastId;
-}
-
-function buildProgressUpdate(
-    int $processedEntries,
-    int $totalEntries,
-    int $totalAdded,
-    int $totalSkipped,
-    int $totalErrors
-): array {
-    if ($totalEntries > 0) {
-        $progress = CHANNEL_PROGRESS_START + (int) floor(($processedEntries / $totalEntries) * (CHANNEL_PROGRESS_END - CHANNEL_PROGRESS_START));
-        $progress = min(PROGRESS_MAX, max(CHANNEL_PROGRESS_START, $progress));
-    } else {
-        $progress = PROGRESS_MAX;
-    }
-
-    $message = sprintf(
-        'Processando canais (%s/%s)...',
-        formatBrazilianNumber($processedEntries),
-        formatBrazilianNumber($totalEntries)
-    );
-
-    return [
-        'progress' => $progress,
-        'message' => $message,
-        'total_added' => $totalAdded,
-        'total_skipped' => $totalSkipped,
-        'total_errors' => $totalErrors,
-    ];
 }
 
 function processJob(PDO $adminPdo, array $job, int $streamTimeout): array
